告警清理 Signed-off-by:wenlong12 <wenlong12@huawei.com>

+++ device/plugins/native_daemon/src/hook_manager.cpp
@@ -163,7 +163,7 @@ bool HookManager::UnloadPlugin(const uint32_t pluginId)
 void HookManager::GetClientConfig(const NativeHookConfig& nativeHookConfig, ClientConfig& clientConfig)
 {
     clientConfig.shareMemroySize = hookConfig_.smb_pages() * PAGE_BYTES;
-    clientConfig.filterSize = hookConfig_.filter_size();
+    clientConfig.filterSize = static_cast<uint32_t>(hookConfig_.filter_size());
     clientConfig.clockId = COMMON::GetClockId(hookConfig_.clock());
     clientConfig.maxStackDepth = hookConfig_.max_stack_depth();
     clientConfig.mallocDisable = hookConfig_.malloc_disable();
